requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.0
Django==5.2.5
django-allauth==65.11.2
idna==3.10
oauthlib==3.3.1
pillow==12.0.0
psycopg2==2.9.10
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
